+++ ChangeLog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 
 ## Release 2.7
-
+- FIX : Compat v21 *07/01/2025* - 2.7.1
 - FIX : Compat v20
   Changed Dolibarr compatibility range to 16 min - 20 max - *04/08/2024* - 2.7.0